Integration (#25)
Updated code in almost every fragment to integrate queries and mutations
for backend. In each ConfirmEntry fragment, added mutations to send the
data entry to the backend. Updated ConfigurationFragment to create
plants and operators. Had to make a slight update to the schema, so
amplify API is regenerated as well.