Frontend touchups and records integration (#31)
In this commit I mostly prepared the app for deployment. I updated the
design of the home page, took away the graph for now, and updated the
design for some of the login/signup fragments. I also ensured there were
no hardcoded strings so that everything is translatable. I finished the
records page integration with the backend.

I made a couple of temporary changes to hide/disable what is still in
progress:
1. I removed the edit button for the records entries, as this feature
still needs work, and
2. I replaced the graphs page with a simple fragment that has an "under
construction" message and icon.
The purpose of these changes was to send out a pre-release version that
Monroe can test while I finish up the graphs and records editing
features.