Set API key expiration to 365 days, amplify pull (#32)
We recently found out that the AguaDatos client Android app was not
synchronizing changes to the AppSync API via Amplify DataStore. This was
because the AppSync API key had expired, and was set to only last for 7
days.

This PR has the following changes:
1. Configure the API key expiration time to 365 days. I did this using
the `amplify api update` command and following the interactive steps.
2. `amplify pull` latest code model changes.

Note that I didn't run `amplify push`, and that we may need to manually
update the API key in `amplifyconfiguration.json`.